added error handling in response
